+++ lib/scss_lint/linter/shorthand.rb
@@ -7,10 +7,13 @@ module SCSSLint
       return unless SHORTHANDABLE_PROPERTIES.include?(property_name)
 
       case node.value
-      when Sass::Script::List
+      when Sass::Script::Tree::Literal
+        # HACK: node_parent may not be initialized at this point, so we need to
+        # set it ourselves
+        node.value.value.node_parent = node.value
+        check_script_string(property_name, node.value.value)
+      when Sass::Script::Tree::ListLiteral
         check_script_list(property_name, node.value)
-      when Sass::Script::String
-        check_script_string(property_name, node.value)
       end
     end